backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
smartsheet-python-sdk==3.0.4
anthropic>=0.28.0
python-dotenv>=1.0.1
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0
python-multipart>=0.0.9
redis>=5.0.4
aioredis>=2.0.1
apscheduler>=3.10.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=24.2.0
mcp>=1.0.0
websockets>=12.0